Implement (and test) filesystem::lockfile
This implements flock(2) file locking via dedicated FDs.  The general
idea here is that I sometimes want to manage my locks seperately from
the file objects that will be used to access that data.  The locks are
RAII-style, so they unlock themselves.

The test cases are probably somewhat fragile, since they rely on
timeouts to check if the locks actually do anything.